requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
crispy-bootstrap4==2022.1
decorator==5.1.1
Django==4.2.5
django-crispy-forms==2.0
executing==1.2.0
ipython==8.15.0
jedi==0.19.0
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dotenv==1.0.0
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.10.1
wcwidth==0.2.6
